split views.py into multiple files
